requirements.txt
python-dotenv
pytest
pylint
psycopg2
confluent-kafka